old_dynarec: Allow building the old dynarec on x64 with VisualStudio2013

+++ src/r4300/x86_64/gr4300.c
@@ -313,7 +313,7 @@ void gendebug(void)
 }
 #endif
 
-void gencallinterp(unsigned long addr, int jump)
+void gencallinterp(unsigned long long addr, int jump)
 {
    free_registers_move_start();
